Modify: Divide layer
