+++ app/src/main/java/kr/hs/pandabear/recom/RecomApplication.kt
@@ -1,0 +1,7 @@
+package kr.hs.pandabear.recom
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class RecomApplication : Application()
