+++ app/src/main/java/kr/hs/pandabear/recom/data/utils/Constants.kt
@@ -1,0 +1,5 @@
+package kr.hs.pandabear.recom.data.utils
+
+object Constants {
+    const val BASE_URL = "http://192.168.1.12:8080/"
+}
