+++ app/src/main/java/kr/hs/pandabear/recom/domain/model/document/Document.kt
@@ -1,0 +1,5 @@
+package kr.hs.pandabear.recom.domain.model.document
+
+data class Document(
+    val code: String
+)
