+++ app/src/main/java/kr/hs/pandabear/recom/domain/model/speech/SpeechInfo.kt
@@ -1,0 +1,6 @@
+package kr.hs.pandabear.recom.domain.model.speech
+
+data class SpeechInfo (
+    val speech: String,
+    var isImpact: Boolean = false
+)
