+++ app/src/main/java/kr/hs/pandabear/recom/domain/request/SaveDataRequest.kt
@@ -1,0 +1,7 @@
+package kr.hs.pandabear.recom.domain.request
+
+data class SaveDataRequest(
+    val content: String,
+    val address: String,
+    val title: String
+)
