--- app/src/main/java/kr/hs/pandabear/recom/network/model/speech/Document.kt
@@ -1,5 +1,0 @@
-package kr.hs.pandabear.recom.network.model.speech
-
-data class Document(
-    val code: String
-)
