--- app/src/main/java/kr/hs/pandabear/recom/network/model/speech/SpeechInfo.kt
@@ -1,6 +1,0 @@
-package kr.hs.pandabear.recom.network.model.speech
-
-data class SpeechInfo (
-    val speech: String,
-    var isImpact: Boolean = false
-)
