--- app/src/main/java/kr/hs/pandabear/recom/network/request/SaveDataRequest.kt
@@ -1,7 +1,0 @@
-package kr.hs.pandabear.recom.network.request
-
-data class SaveDataRequest(
-    val content: String,
-    val address: String,
-    val title: String
-)
