--- app/src/main/java/kr/hs/pandabear/recom/network/response/Response.kt
@@ -1,7 +1,0 @@
-package kr.hs.pandabear.recom.network.response
-
-data class Response<T>(
-    val `data`: T,
-    val status: Int,
-    val message: String
-)
