+++ app/src/main/java/kr/hs/pandabear/recom/presentation/widget/Resource.kt
@@ -1,0 +1,7 @@
+package kr.hs.pandabear.recom.presentation.widget
+
+sealed class Resource<T>(val data: T? = null, val message: String? = null) {
+    class Success<T>(data: T): Resource<T>(data)
+    class Error<T>(message: String, data: T? = null): Resource<T>(data, message)
+    class Loading<T>(data: T? = null) : Resource<T>(data)
+}
