+++ app/src/main/java/kr/hs/pandabear/recom/presentation/widget/view/SoundVisualizerView.kt
@@ -1,0 +1,104 @@
+package kr.hs.pandabear.recom.presentation.widget.view
+
+import android.content.Context
+import android.graphics.Canvas
+import android.graphics.Paint
+import android.util.AttributeSet
+import android.view.View
+import kr.hs.pandabear.recom.R
+
+class SoundVisualizerView(
+    context: Context,
+    attrs: AttributeSet? = null
+) : View(context, attrs) {
+
+    var onRequestCurrentAmplitude: (() -> Int)? = null
+
+    private val amplitudePaint = Paint(Paint.ANTI_ALIAS_FLAG).apply {
+        color = context.getColor(R.color.color_main)
+        strokeWidth = LINE_WIDTH
+        strokeCap = Paint.Cap.ROUND
+    }
+
+    private var drawingWidth: Int = 0
+    private var drawingHeight: Int = 0
+    private var drawingAmplitudes: List<Int> = emptyList()
+    private var isReplaying: Boolean = false
+    private var replayingPosition: Int = 0
+
+    private val visualizerRepeatAction: Runnable = object : Runnable {
+        override fun run() {
+            if (!isReplaying) {
+                val currentAmplitude = onRequestCurrentAmplitude?.invoke() ?: 0
+                drawingAmplitudes = listOf(currentAmplitude) + drawingAmplitudes
+            } else {
+                replayingPosition++
+            }
+            invalidate()
+
+            handler?.postDelayed(this, ACTION_INTERVAL)
+        }
+    }
+
+    override fun onSizeChanged(w: Int, h: Int, oldw: Int, oldh: Int) {
+        super.onSizeChanged(w, h, oldw, oldh)
+
+        drawingWidth = w
+        drawingHeight = h
+    }
+
+    override fun onDraw(canvas: Canvas?) {
+        super.onDraw(canvas)
+
+        canvas ?: return
+
+        val centerY = drawingHeight / 2f
+        var offsetX = drawingWidth.toFloat()
+
+        drawingAmplitudes
+            .let { amplitudes ->
+                if (isReplaying) {
+                    amplitudes.takeLast(replayingPosition)
+                } else {
+                    amplitudes
+                }
+            }
+            .forEach { amplitude ->
+                val lineLength = amplitude / MAX_AMPLITUDE * drawingHeight * 0.8F
+
+                offsetX -= LINE_SPACE
+                if (offsetX < 0) return@forEach
+
+                canvas.drawLine(
+                    offsetX,
+                    centerY - lineLength / 1.3F,
+                    offsetX,
+                    centerY + lineLength / 1.3F,
+                    amplitudePaint
+                )
+            }
+    }
+
+    fun startVisualizing(isReplaying: Boolean) {
+        this.isReplaying = isReplaying
+        handler?.post(visualizerRepeatAction)
+    }
+
+    fun stopVisualizing() {
+        replayingPosition = 0
+        handler?.removeCallbacks(visualizerRepeatAction)
+    }
+
+    fun clearVisualization() {
+        drawingAmplitudes = emptyList()
+        invalidate()
+    }
+
+    companion object {
+        private const val LINE_WIDTH = 10F
+        private const val LINE_SPACE = 15F
+        private const val MAX_AMPLITUDE = Short.MAX_VALUE.toFloat()
+        private const val ACTION_INTERVAL = 20L
+    }
+
+}
