--- app/src/main/java/kr/hs/pandabear/recom/view/base/BaseActivity.kt
@@ -1,77 +1,0 @@
-package kr.hs.pandabear.recom.view.base
-
-import android.os.Bundle
-import android.widget.Toast
-import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-import kr.hs.pandabear.recom.BR
-import kr.hs.pandabear.recom.R
-import kr.hs.pandabear.recom.viewmodel.base.BaseViewModel
-import java.lang.reflect.ParameterizedType
-import java.util.*
-
-abstract class BaseActivity<VB : ViewDataBinding, VM: BaseViewModel> : AppCompatActivity() {
-    protected lateinit var mBinding: VB
-    protected lateinit var mViewModel: VM
-
-    protected abstract val viewModel: VM
-
-    protected abstract fun observerViewModel()
-    protected abstract fun bindingViewEvent()
-
-    protected open fun onErrorEvent(e: Throwable) {
-        Toast.makeText(this, e.toString(), Toast.LENGTH_SHORT).show()
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        performDataBinding()
-        observerViewModel()
-        bindingViewEvent()
-
-        mViewModel.onErrorEvent.observe(this) {
-            onErrorEvent(it)
-        }
-    }
-
-    private fun performDataBinding() {
-        mBinding = DataBindingUtil.setContentView(this, layoutRes())
-        mViewModel = if (::mViewModel.isInitialized) mViewModel else viewModel
-        mBinding.setVariable(BR.vm, mViewModel)
-        mBinding.lifecycleOwner = this
-        mBinding.executePendingBindings()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        if(::mBinding.isInitialized) mBinding.unbind()
-    }
-
-    @LayoutRes
-    private fun layoutRes(): Int {
-        val split = ((Objects.requireNonNull(javaClass.genericSuperclass) as ParameterizedType).actualTypeArguments[0] as Class<*>)
-            .simpleName.replace("Binding$".toRegex(), "").split("(?<=.)(?=\\p{Upper})".toRegex())
-            .dropLastWhile { it.isEmpty() }.toTypedArray()
-
-        val name = StringBuilder()
-
-        for (i in split.indices) {
-            name.append(split[i].lowercase(Locale.ROOT))
-            if (i != split.size - 1)
-                name.append("_")
-        }
-
-        try {
-            return R.layout::class.java.getField(name.toString()).getInt(R.layout::class.java)
-        } catch (e: IllegalAccessException) {
-            e.printStackTrace()
-        } catch (e: NoSuchFieldException) {
-            e.printStackTrace()
-        }
-
-        return 0
-    }
-}
