--- app/src/main/java/kr/hs/pandabear/recom/widget/Constants.kt
@@ -1,5 +1,0 @@
-package kr.hs.pandabear.recom.widget
-
-object Constants {
-    const val BASE_URL = "http://192.168.1.12:8080/"
-}
