--- app/src/main/java/kr/hs/pandabear/recom/widget/RecomApplication.kt
@@ -1,7 +1,0 @@
-package kr.hs.pandabear.recom.widget
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class RecomApplication : Application()
