--- app/src/main/java/kr/hs/pandabear/recom/widget/Resource.kt
@@ -1,7 +1,0 @@
-package kr.hs.pandabear.recom.widget
-
-sealed class Resource<T>(val data: T? = null, val message: String? = null) {
-    class Success<T>(data: T): Resource<T>(data)
-    class Error<T>(message: String, data: T? = null): Resource<T>(data, message)
-    class Loading<T>(data: T? = null) : Resource<T>(data)
-}
